// src/profiles/types/enums.ts
export enum DanceStyle {
    BALLET = 'Ballet',
    HIP_HOP = 'Hip Hop',
    SALSA = 'Salsa',
    CONTEMPORARY = 'Contemporary',
    JAZZ = 'Jazz',
    TAP = 'Tap',
  }
  
  export enum Goal {
    IMPROVE_TECHNIQUE = 'Improve Technique',
    WIN_COMPETITIONS = 'Win Competitions',
    LEARN_NEW_STYLES = 'Learn New Styles',
    MAKE_FRIENDS = 'Make Friends',
    FITNESS = 'Fitness',
  }
  
  export enum DanceLevel {
    BEGINNER = 'Beginner',
    INTERMEDIATE = 'Intermediate',
    ADVANCED = 'Advanced',
    PROFESSIONAL = 'Professional',
  }
  
  export enum Role {
    DANCER = 'Dancer',
    CHOREOGRAPHER = 'Choreographer',
    TEACHER = 'Teacher',
    STUDENT = 'Student',
  }